Move orderbook and actions cells to separate files
