+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/LoansTokenActiveTable.module.less
@@ -37,6 +37,14 @@
 // * Table *//
 
 // * Cells *//
+.actionsButtons {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 8px;
+  width: 100%;
+}
+
 .headerTitleRow {
   display: flex;
   align-items: center;
