--- src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/ActionsCell/OrderBook/OrderBook.module.less
@@ -1,56 +1,0 @@
-.container {
-  padding-top: 12px;
-
-  width: 100%;
-  height: 100%;
-}
-
-.table {
-  width: 100%;
-
-  tr > td,
-  tr > th {
-    height: 44px;
-    width: 100%;
-  }
-}
-
-.tableWrapper {
-  height: 200px;
-}
-
-// * Cells * //
-.checkboxRow {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.checkbox {
-  height: 16px;
-  padding-left: 18px;
-}
-
-.aprRow {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.aprValue {
-  text-align: center;
-}
-
-.borrowValueContainer {
-  display: flex;
-  flex-direction: column;
-
-  div {
-    font: var(--important-text-lg);
-  }
-}
-
-.ltvValue {
-  font: var(--body-text-md);
-}
-// * Cells * //
