--- src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/ActionsCell/index.ts
@@ -1,1 +1,0 @@
-export * from './ActionsCell'
