+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/RefinanceTokenModal/OrderBook/OrderBook.tsx
@@ -1,5 +1,6 @@
-import { FC, useCallback, useMemo } from 'react'
+import { FC } from 'react'
 
+import { BN } from 'fbonds-core'
 import { BondOfferV3 } from 'fbonds-core/lib/fbond-protocol/types'
 
 import Table from '@banx/components/Table'
@@ -8,7 +9,6 @@ import { core } from '@banx/api/tokens'
 import { useNftTokenType } from '@banx/store/nft'
 
 import { getTableColumns } from './columns'
-import { useSelectedOffer } from './useSelectedOffers'
 
 import styles from './OrderBook.module.less'
 
@@ -16,37 +16,19 @@ interface OrderBookProps {
   offers: BondOfferV3[]
   loan: core.TokenLoan
   isLoading: boolean
+  refinance: (offer: BondOfferV3, tokensToRefinance: BN) => void
 }
 
-const OrderBook: FC<OrderBookProps> = ({ loan, offers, isLoading }) => {
+const OrderBook: FC<OrderBookProps> = ({ loan, offers, isLoading, refinance }) => {
   const { tokenType } = useNftTokenType()
 
-  const { toggle: toggleOffer, find: findOfferInSelection } = useSelectedOffer()
-
-  const onRowClick = useCallback(
-    (offer: BondOfferV3) => {
-      toggleOffer(offer)
-    },
-    [toggleOffer],
-  )
-
-  const columns = getTableColumns({
-    findOfferInSelection,
-    onRowClick,
-    tokenType,
-    loan,
-  })
-
-  const rowParams = useMemo(() => {
-    return { onRowClick }
-  }, [onRowClick])
+  const columns = getTableColumns({ refinance, tokenType, loan })
 
   return (
     <div className={styles.container}>
       <Table
         data={offers}
         columns={columns}
-        rowParams={rowParams}
         className={styles.table}
         classNameTableWrapper={styles.tableWrapper}
         loading={isLoading}
