+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/index.ts
@@ -1,2 +1,1 @@
-export * from './ActionsCell'
 export * from './cells'
