+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/helpers.ts
@@ -1,15 +1,22 @@
+import { BN } from 'fbonds-core'
 import {
   calculateCurrentInterestSolPure,
   calculatePartOfLoanBodyFromInterest,
 } from 'fbonds-core/lib/fbond-protocol/functions/perpetual'
+import { BondOfferV3 } from 'fbonds-core/lib/fbond-protocol/types'
 import { map } from 'lodash'
 import moment from 'moment'
 
+import { convertBondOfferV3ToCore } from '@banx/api/nft'
 import { core } from '@banx/api/tokens'
 import { BONDS } from '@banx/constants'
 import {
+  ZERO_BN,
+  adjustAmountWithUpfrontFee,
   caclulateBorrowTokenLoanValue,
   calcWeightedAverage,
+  calculateIdleFundsInOffer,
+  calculateTokenLoanValueWithUpfrontFee,
   isBanxSolTokenType,
   isTokenLoanRepaymentCallActive,
 } from '@banx/utils'
@@ -85,4 +92,42 @@ export const calcWeightedApr = (loans: core.TokenLoan[]) => {
 
   const totalRepayValues = map(loans, (loan) => caclulateBorrowTokenLoanValue(loan).toNumber())
   return calcWeightedAverage(totalAprValues, totalRepayValues)
+}
+
+type CalculateTokenToGet = (props: {
+  offer: BondOfferV3
+  loan: core.TokenLoan
+  marketTokenDecimals: number
+}) => BN
+
+export const calculateTokenToGet: CalculateTokenToGet = ({ offer, loan, marketTokenDecimals }) => {
+  const maxTokenToGet = calculateIdleFundsInOffer(convertBondOfferV3ToCore(offer))
+
+  const tokenSupply = loan.fraktBond.fbondTokenSupply
+  const collateralsPerToken = offer.validation.collateralsPerToken
+
+  if (!tokenSupply || !collateralsPerToken) return ZERO_BN
+
+  const marketTokenDecimalsMultiplier = new BN(10).pow(new BN(marketTokenDecimals))
+
+  const tokenToGet = BN.min(
+    new BN(tokenSupply).mul(marketTokenDecimalsMultiplier).div(collateralsPerToken),
+    maxTokenToGet,
+  )
+
+  const adjustedTokenToGet = adjustAmountWithUpfrontFee(tokenToGet)
+
+  return adjustedTokenToGet
+}
+
+export const getCurrentLoanInfo = (loan: core.TokenLoan) => {
+  const currentLoanDebt = caclulateBorrowTokenLoanValue(loan).toNumber()
+  const currentLoanBorrowedAmount = calculateTokenLoanValueWithUpfrontFee(loan).toNumber()
+  const currentApr = loan.bondTradeTransaction.amountOfBonds
+
+  return {
+    currentLoanDebt,
+    currentLoanBorrowedAmount,
+    currentApr,
+  }
 }
